Adds REST endpoints for sequence voting
Adds REST endpoints to handle sequence requests and votes, integrating with the GraphQL mutation service.
Introduces request and response objects for the new endpoints, and includes exception handling.
Also, improves the check if a user has already voted or requested a sequence by not checking the IP address if it is empty.

Updates exception handling to use custom exceptions to return error messages to the client.